Base project
The algorithm works properly in isolation in the Modules.py file. With this in mind, both modules, seeking_algorithm and tracing_algorithm, have been implemented in the main algorithm script CSO_manual.

However, errors are still generated in the CSO_manual.py file for some reason. I will check the attributes within the fs() function to fix the errors and make a commit with the first functional version of the CSO Algorithm.